Allow hiding tab buttons in sidebars
Sometimes when you enable a plugin you don't want its whole toolview
to be around all the time snippets for e.g., is such a plugin whose
toolview is rarely a used thing for me. However, the toolview and its
respective button in the sidebar is always there increasing clutter.

This change introduces a feature which allows a user to hide whatever
buttons they want. This allows kate to be even more minimal.

I haven't coupled "tool view showing" with the "button show/hide" as
both can be used independent of each other. For e.g., you can hide the
"Documents" button in the sidebar but still use a shortcut to see the
toolview. This will allow people to even completely hide the sidebars
and still be able to use all the plugin features and toolviews on demand

One more thing that can be done as a follow up is to create actions
that hide a sidebar completely.